requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
httpx
motor  
passlib[bcrypt] 
python-jose
pydantic[email]